aniadido: tags para articulos. Descuentos. Creando ventas

+++ database/migrations/2023_10_31_025152_create_articulo_calzado.php
@@ -16,9 +16,8 @@ return new class extends Migration
             $table->unsignedBigInteger('articulo_id');
             $table->unsignedBigInteger('calzado_id');
             $table->integer('stocks');
-            $table->integer('precio');
+            $table->decimal('precio', 8, 2);
             $table->timestamps();
-
             $table->foreign('articulo_id')->references('id')->on('articulos')->onDelete('cascade');
             $table->foreign('calzado_id')->references('id')->on('calzados')->onDelete('cascade');
